+++ buildSrc/src/main/kotlin/util/gradle.kt
@@ -8,27 +8,26 @@ import java.nio.charset.Charset
 
 typealias Lambda<R, V> = R.() -> V
 
-val CI by lazy { "true".equals(System.getenv("CI"), true) }
+val CI by lazy { !"false".equals(System.getenv("CI") ?: "false", true) }
+val SANDBOX by lazy { !"false".equals(System.getenv("SANDBOX") ?: "false", true) }
 
-fun <R, V> Lambda<R, V>.toClosure(owner: Any? = null, thisObj: Any? = null) = object : Closure<V>(owner, thisObj) {
-  @Suppress("UNCHECKED_CAST")
-  fun doCall() {
-    with(delegate as R) {
-      this@toClosure()
+fun <R, V> Lambda<R, V>.toClosure(owner: Any? = null, thisObj: Any? = null) =
+    object : Closure<V>(owner, thisObj) {
+      @Suppress("UNCHECKED_CAST")
+      fun doCall() {
+        with(delegate as R) { this@toClosure() }
+      }
     }
-  }
-}
 
-fun <R, V> closureOf(owner: Any? = null, thisObj: Any? = null, func: R.() -> V) = func.toClosure(owner, thisObj)
+fun <R, V> closureOf(owner: Any? = null, thisObj: Any? = null, func: R.() -> V) =
+    func.toClosure(owner, thisObj)
 
 infix fun <T> Property<T>.by(value: T) {
   set(value)
 }
 
 object Git {
-  val headCommitHash by lazy {
-    execAndCapture("git rev-parse --verify HEAD")
-  }
+  val headCommitHash by lazy { execAndCapture("git rev-parse --verify HEAD") }
 }
 
 fun execAndCapture(cmd: String): String? {
@@ -39,4 +38,5 @@ fun execAndCapture(cmd: String): String? {
   } else null
 }
 
-val Project.isMainHost: Boolean get() = HostManager.simpleOsName().equals("${properties["project.mainOS"]}", true)
+val Project.isMainHost: Boolean
+  get() = HostManager.simpleOsName().equals("${properties["project.mainOS"]}", true)