+++ lib/template-kmp-library-core/src/commonMain/kotlin/CoreLib.kt
@@ -5,4 +5,5 @@ internal expect val platform: String
 public class CoreLib {
   public fun sampleApi(): String = platform
   public suspend fun sampleSuspendApi(): String = sampleApi()
+  public val sampleValue: Number = 420 * 69
 }